[REFACTOR] 1. Delete PostLayout
