+++ frontend/src/containers/post/PostDetail.tsx
@@ -5,7 +5,7 @@ import { useOnClickOutside } from 'usehooks-ts';
 import styled from 'styled-components';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faThumbsDown, faThumbsUp } from '@fortawesome/free-regular-svg-icons';
-import { faStar } from '@fortawesome/free-solid-svg-icons';
+import { faSearch, faStar } from '@fortawesome/free-solid-svg-icons';
 import { RootState } from 'index';
 import { timeAgoFormat } from 'utils/datetime';
 import { funcTypeToStr, postActions } from 'store/slices/post';
@@ -15,7 +15,6 @@ import { BlueBigBtn, CommentGreenBtn, RedSmallBtn, GreenCommentSubmitBtn } from 
 import { TagBubble } from 'components/tag/tagbubble';
 import { ColumnCenterFlex, ColumnFlex, RowCenterFlex } from 'components/post/layout';
 import { UserDetailHorizontalModal, UserDetailModal } from 'components/post/UserDetailModal';
-import { PostDetailLayout, PostPageWrapper, SideBarWrapper } from './PostLayout';
 import ImageDetailModal from 'components/post/ImageDetailModal';
 
 export interface IPropsComment {
@@ -28,6 +27,10 @@ export const FuncType = {
   Dislike: 'dislike',
   Scrap: 'scrap',
 };
+
+interface IPropsSearchClear {
+  isActive?: boolean;
+}
 
 interface IPropsFuncBtn {
   color: string;
@@ -566,9 +569,56 @@ const PostDetail = () => {
       <SideBarItem>사이드바 공간2</SideBarItem>
     </SideBarWrapper>
   );
+
+  const postSearch = useSelector(({ post }: RootState) => post.postSearch);
+  const [search, setSearch] = useState(postSearch);
+
+  useEffect(() => {
+    setSearch(postSearch);
+  }, []);
   return (
     <PostPageWrapper>
-      {PostDetailLayout(PostDetailContent, SideBar)}
+      <PostContentWrapper>
+        <TopWrapper>
+          <SearchForm
+            onSubmit={e => {
+              e.preventDefault();
+              navigate(`/post`);
+              dispatch(
+                postActions.postSearch({
+                  search_keyword: search,
+                }),
+              );
+            }}
+          >
+            <SearchIcon>
+              <FontAwesomeIcon icon={faSearch} />
+            </SearchIcon>
+            <SearchInput
+              placeholder="Search keyword"
+              value={search}
+              onChange={e => setSearch(e.target.value)}
+            ></SearchInput>
+            <ClearSearchInput
+              isActive={search !== ''}
+              onClick={() => {
+                setSearch('');
+                dispatch(
+                  postActions.postSearch({
+                    search_keyword: '',
+                  }),
+                );
+              }}
+            >
+              Clear
+            </ClearSearchInput>
+          </SearchForm>
+        </TopWrapper>
+        <Main_SideWrapper>
+          {PostDetailContent}
+          {SideBar}
+        </Main_SideWrapper>
+      </PostContentWrapper>
       {ImageDetailModal({
         isActive: imageModalOpen,
         onClose: () => setImageModalOpen(false),
@@ -904,6 +954,76 @@ const PostUploadedImage = styled.img`
   border-radius: 15px;
   object-fit: cover;
   cursor: pointer;
+`;
+
+const SearchForm = styled.form`
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+`;
+
+const SearchInput = styled.input`
+  width: 95%;
+  padding: 15px 20px;
+  font-size: 15px;
+  border: none;
+`;
+
+const ClearSearchInput = styled.span<IPropsSearchClear>`
+  width: 5%;
+  text-align: center;
+  cursor: pointer;
+  ${({ isActive }) =>
+    !isActive &&
+    `
+    display: none;
+  `}
+`;
+
+const SearchIcon = styled(RowCenterFlex)`
+  margin-left: 20px;
+`;
+
+export const PostPageWrapper = styled(ColumnCenterFlex)`
+  background-color: var(--fit-green-back);
+  width: 100%;
+  height: 100%;
+  min-height: 100vh;
+  overflow-x: hidden;
+  position: relative;
+`;
+
+export const PostContentWrapper = styled(ColumnCenterFlex)`
+  width: 100%;
+  height: 100%;
+  min-height: 100vh;
+  max-width: 1200px;
+
+  @media all and (max-width: 650px) {
+    width: 100%;
+  }
+`;
+
+export const TopWrapper = styled.div`
+  margin: 40px 0px 15px 0px;
+  width: 100%;
+  background-color: var(--fit-white);
+`;
+
+export const Main_SideWrapper = styled.div`
+  display: grid;
+  grid-template-columns: 8fr 2fr;
+  row-gap: 10px;
+  column-gap: 10px;
+  width: 100%;
+  height: 80vh;
+  min-height: 640px;
+  margin-bottom: 50px;
+`;
+
+export const SideBarWrapper = styled.div`
+  width: 100%;
 `;
 
 export default PostDetail;
